+++ src/utils/process.js
@@ -1,4 +1,5 @@
 import Process from 'child_process'
+import fs from 'fs'
 
 const os = process.platform
 //TODO: win32
@@ -12,7 +13,7 @@ const options = {
     env: null
 };
 
-const doCommand = async function ( command ){
+const doCommand = async ( command ) => {
   // 捕获标准输出并将其打印到控制台
   return new Promise((resolve, reject) => {
     Process.exec(command, options, (e, stdout, stderr) => {
@@ -24,4 +25,19 @@ const doCommand = async function ( command ){
     })
   })
 }
-export { doCommand }
+
+const execShell = async ( shellPath, params ) => {
+  if(!fs.existsSync(shellPath)){
+    return Promise.reject({ message: `shell script file not found ${shellPath}`})
+  }
+  return new Promise((resolve, reject) => {
+    Process.execFile(shellPath, params || [], options, (e, stdout, stderr) => {
+      if(e){
+        reject( {data: stderr })
+      }else{
+        resolve( {data: stdout })
+      }
+    })
+  })
+}
+export { doCommand, execShell }